ltm: strip element subscripts in discovery exit-port recompute
Discovery runs on the element-level graph, so an arrayed loop's non-module
nodes carry element subscripts (s[nyc] -> m -> growth[nyc]).
recompute_module_input_edge_series matched the entry port with an exact
`normalize_module_ref(inp.src) == link.from` and looked the exit reader up in
the bare-keyed variables() map -- both fail on the subscripted forms, so the
recompute declined for EVERY arrayed module loop and the wrong-exit-port
composite bug the override exists to fix re-occurred. Strip the subscript via
the canonical crate::ltm::strip_subscript on link.from (entry match),
link.to/next.from (module node + sequentiality guard), and next.to (exit
reader lookup) before each name comparison, preserving the prior ambiguity
semantics (two distinct ports -> decline; same port twice -> unique). This
mirrors the exhaustive twin compute_module_link_overrides, which already
strips link.from/link.to/next.from/next.to (db/ltm/mod.rs).

The end-to-end arrayed manifestation is currently masked by an upstream gap:
a scalar module output feeding an arrayed reader (growth[Region] = m.pos)
emits a single scalar constant-0 link score rather than per-element
m->growth[e] scores, so the loop product is 0 and discovery drops the loop
before the recompute runs. That gap is tracked separately as GH #716. This
change is therefore latent defense that restores name-handling parity with
the exhaustive override; it is covered by a unit-level red-green test
(recompute_strips_element_subscripts_before_port_match) that drives the
matching code directly with a real compiled module project's graph/results
and a hand-built subscripted links chain (fails returning None before, passes
returning the pos-pathway series after), plus a documentation test
(analyze_model_arrayed_module_loop_blocked_by_scalar_output_gap) pinning the
current masked end-to-end state so closing #716 is a deliberate change.

The `ltm_finding.rs` test module moved to a sibling `ltm_finding_tests.rs`
(mounted via `#[cfg(test)] #[path]`) to keep the production file under the
per-file line cap.

Reference: https://github.com/bpowers/simlin/pull/705#discussion_r3353758167
GH #698
